+++ src/main/java/com/kustacks/kuring/admin/common/dto/response/CategoryDto.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.common.dto;
+package com.kustacks.kuring.admin.common.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,10 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class HeartBeatResponseDto {
+public class CategoryDto {
 
-    private String type;
-
-    private String content;
+    private String name;
 }
-
