+++ src/main/java/com/kustacks/kuring/admin/presentation/AdminController.java
@@ -8,9 +8,9 @@ import com.kustacks.kuring.category.business.CategoryService;
 import com.kustacks.kuring.category.domain.Category;
 import com.kustacks.kuring.common.annotation.CheckSession;
 import com.kustacks.kuring.common.dto.AdminMessageDto;
-import com.kustacks.kuring.common.dto.CategoryDto;
-import com.kustacks.kuring.common.dto.FakeUpdateResponseDto;
-import com.kustacks.kuring.common.dto.LoginResponseDto;
+import com.kustacks.kuring.admin.common.dto.response.CategoryDto;
+import com.kustacks.kuring.admin.common.dto.response.FakeUpdateResponseDto;
+import com.kustacks.kuring.admin.common.dto.response.LoginResponseDto;
 import com.kustacks.kuring.common.dto.NoticeMessageDto;
 import com.kustacks.kuring.common.dto.ResponseDto;
 import com.kustacks.kuring.common.error.APIException;
