+++ src/main/java/com/kustacks/kuring/category/presentation/CategoryController.java
@@ -3,9 +3,9 @@ package com.kustacks.kuring.category.presentation;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.kustacks.kuring.category.business.CategoryService;
 import com.kustacks.kuring.common.firebase.FirebaseService;
-import com.kustacks.kuring.common.dto.CategoryListResponseDto;
-import com.kustacks.kuring.common.dto.SubscribeCategoriesRequestDto;
-import com.kustacks.kuring.common.dto.SubscribeCategoriesResponseDto;
+import com.kustacks.kuring.category.common.dto.response.CategoryListResponseDto;
+import com.kustacks.kuring.category.common.dto.request.SubscribeCategoriesRequestDto;
+import com.kustacks.kuring.category.common.dto.response.SubscribeCategoriesResponseDto;
 import com.kustacks.kuring.category.domain.Category;
 import com.kustacks.kuring.user.business.UserService;
 import com.kustacks.kuring.user.domain.User;
