--- src/main/java/com/kustacks/kuring/common/dto/FcmMessageDto.java
@@ -1,6 +1,0 @@
-package com.kustacks.kuring.common.dto;
-
-public class FcmMessageDto {
-
-    protected String type;
-}
