+++ src/main/java/com/kustacks/kuring/feedback/common/dto/request/SaveFeedbackRequestDto.java
@@ -1,19 +1,17 @@
-package com.kustacks.kuring.common.dto;
+package com.kustacks.kuring.feedback.common.dto.request;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
-
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SubscribeCategoriesRequestDto {
+public class SaveFeedbackRequestDto {
 
     private String id;
 
-    private List<String> categories;
+    private String content;
 }
+
