+++ src/main/java/com/kustacks/kuring/notice/common/dto/response/NoticeResponseDto.java
@@ -1,5 +1,6 @@
-package com.kustacks.kuring.common.dto;
+package com.kustacks.kuring.notice.common.dto.response;
 
+import com.kustacks.kuring.common.dto.ResponseDto;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
