+++ src/main/java/com/kustacks/kuring/search/common/dto/response/HeartBeatResponseDto.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.common.dto;
+package com.kustacks.kuring.search.common.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SaveFeedbackRequestDto {
+public class HeartBeatResponseDto {
 
-    private String id;
+    private String type;
 
     private String content;
 }
