+++ src/main/java/com/kustacks/kuring/search/presentation/SearchHandler.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.search.handler;
+package com.kustacks.kuring.search.presentation;
 
 import org.springframework.web.socket.WebSocketSession;
