+++ src/main/java/com/kustacks/kuring/search/presentation/StaffWebSocketHandler.java
@@ -1,15 +1,14 @@
-package com.kustacks.kuring.search.handler;
+package com.kustacks.kuring.search.presentation;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kustacks.kuring.common.dto.NoticeSearchDto;
-import com.kustacks.kuring.common.dto.NoticeWebSocketResponseDto;
-import com.kustacks.kuring.notice.domain.Notice;
+import com.kustacks.kuring.common.dto.StaffDto;
+import com.kustacks.kuring.search.common.dto.response.StaffWebSocketResponseDto;
+import com.kustacks.kuring.staff.domain.Staff;
 import com.kustacks.kuring.common.error.ErrorCode;
 import com.kustacks.kuring.common.error.WebSocketExceptionHandler;
-import com.kustacks.kuring.notice.business.NoticeService;
+import com.kustacks.kuring.staff.business.StaffService;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
@@ -20,27 +19,20 @@ import java.util.List;
 
 @Slf4j
 @Component
-public class NoticeWebSocketHandler implements SearchHandler {
+public class StaffWebSocketHandler implements SearchHandler {
 
-    @Value("${notice.normal-base-url}")
-    private String normalBaseUrl;
+    private final String ERROR_TYPE = "staff";
 
-    @Value("${notice.library-base-url}")
-    private String libraryBaseUrl;
-
-    private final String ERROR_TYPE = "notice";
-
-    private final NoticeService noticeService;
+    private final StaffService staffService;
     private final ObjectMapper objectMapper;
     private final WebSocketExceptionHandler exceptionHandler;
 
-
-    public NoticeWebSocketHandler(
-            NoticeService noticeService,
+    public StaffWebSocketHandler(
+            StaffService staffService,
             ObjectMapper objectMapper,
             WebSocketExceptionHandler exceptionHandler) {
 
-        this.noticeService = noticeService;
+        this.staffService = staffService;
         this.objectMapper = objectMapper;
         this.exceptionHandler = exceptionHandler;
     }
@@ -53,15 +45,14 @@ public class NoticeWebSocketHandler implements SearchHandler {
             return;
         }
 
-        List<Notice> notices = noticeService.handleSearchRequest(keywords);
+        List<Staff> searchResult = staffService.handleSearchRequest(keywords);
 
-        LinkedList<NoticeSearchDto> noticeDTOList = new LinkedList<>();
-        for (Notice notice : notices) {
-            noticeDTOList.add(NoticeSearchDto.entityToDTO(notice, notice.getCategory().getName().equals("library")
-                    ? libraryBaseUrl : normalBaseUrl));
+        List<StaffDto> searchResultDTOList = new LinkedList<>();
+        for (Staff staff : searchResult) {
+            searchResultDTOList.add(StaffDto.entityToDto(staff));
         }
 
-        NoticeWebSocketResponseDto responseObject = new NoticeWebSocketResponseDto(noticeDTOList);
+        StaffWebSocketResponseDto responseObject = new StaffWebSocketResponseDto(searchResultDTOList);
 
         try {
             String responseString = objectMapper.writeValueAsString(responseObject);
@@ -70,7 +61,7 @@ public class NoticeWebSocketHandler implements SearchHandler {
             if(e instanceof JsonProcessingException) {
                 exceptionHandler.sendErrorMessage(session, ErrorCode.WS_CANNOT_STRINGIFY, ERROR_TYPE);
             } else {
-                exceptionHandler.sendErrorMessage(session, ErrorCode.WS_CANNOT_SEND, ERROR_TYPE);
+                exceptionHandler.sendErrorMessage(session, ErrorCode.WS_SERVER_ERROR, ERROR_TYPE);
             }
             log.error("", e);
         }
